requirements.txt
gunicorn==19.10.0
PyMySQL==1.0.2
Flask==0.10.1
Django==3.2.3
django-cors-headers==3.7.0
wheel==0.36.2
google-cloud-storage==1.38.0
scipy==1.7.0

